chore: status.ts -> httpClient.const.ts 이름 변경, status.type.ts -> httpClient.type.ts로 이동 	modified:   src/services/httpClient/__test__/utils.test.ts 	renamed:    src/services/httpClient/status.ts -> src/services/httpClient/httpClient.const.ts 	modified:   src/services/httpClient/httpClient.ts 	modified:   src/services/httpClient/httpClient.type.ts 	deleted:    src/services/httpClient/status.type.ts 	modified:   src/services/httpClient/utils.ts 	modified:   src/services/openWeatherOneCall/axios.ts
